+++ pkg/learn/store_test.go
@@ -216,3 +216,4 @@ func TestStore_UpdateReuseCount_NotFound(t *testing.T) {
 	// Then: error
 	assert.Error(t, err)
 }
+